requirements.txt
streamlit>=1.24.0
snowflake-connector-python>=3.0.0
boto3>=1.28.0
pandas>=2.0.0
pillow>=9.0.0
cachetools>=5.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
altair>=5.0.0